+++ tests/test_tensorstore_serialization.py
@@ -6,7 +6,7 @@ import jax.numpy as jnp
 import numpy as np
 import optax
 from chex import assert_trees_all_close
-from utils import MLP, arrays_only, assert_trees_not_close
+from test_utils import MLP, arrays_only, assert_trees_not_close
 
 from levanter.tensorstore_serialization import tree_deserialize_leaves_tensorstore, tree_serialize_leaves_tensorstore
